Change from float32 to float64

+++ AbvCalculator.go
@@ -2,25 +2,25 @@ package main
 
 
 type AbvRequest struct {
-	OriginalGravity	float32	`json:"originalGravity"`
-	FinalGravity float32 `json:"finalGravity"`
+	OriginalGravity	float64	`json:"originalGravity"`
+	FinalGravity float64 `json:"finalGravity"`
 }
 
 type AbvResponse struct {
-	StandardAbv float32 `json:"standardAbv"`
-	AlternateAbv float32 `json:"alternateAbv"`
-	TotalCalories float32 `json:"totalCalories"`
+	StandardAbv float64 `json:"standardAbv"`
+	AlternateAbv float64 `json:"alternateAbv"`
+	TotalCalories float64 `json:"totalCalories"`
 }
 
-func calculateStandardAbv(originalGravity float32, finalGravity float32) float32 {
+func calculateStandardAbv(originalGravity float64, finalGravity float64) float64 {
 	return (originalGravity - finalGravity) * 131.25
 }
 
-func calculateAlternateAbv(originalGravity float32, finalGravity float32) float32 {
+func calculateAlternateAbv(originalGravity float64, finalGravity float64) float64 {
 	return (76.08 * (originalGravity - finalGravity) / (1.775 - originalGravity)) * (finalGravity / 0.794)
 }
 
-func calculateTotalCalories(originalGravity float32, finalGravity float32) float32 {
+func calculateTotalCalories(originalGravity float64, finalGravity float64) float64 {
 	caloriesFromCarbs := 3500.0 * finalGravity * ((0.1808 * originalGravity) + (0.819 * finalGravity) - 1.0004)
 	caloriesFromAlcohol := 1881.22 * finalGravity * (originalGravity - finalGravity)/(1.775 - originalGravity)
